# Optimize insert statement generate