## Description

Optimize tsbs insert performance.
set wal_mode=DISABLE
Before:
```
Summary:
loaded 78537600 metrics in 75.496sec with 1 workers (mean rate 1040289.76 metrics/sec)
loaded 6998400 rows in 75.496sec with 1 workers (mean rate 92699.09 rows/sec)
```
After
```
Summary:
loaded 78537600 metrics in 29.084sec with 1 workers (mean rate 2700401.36 metrics/sec)
loaded 6998400 rows in 29.084sec with 1 workers (mean rate 240629.82 rows/sec)
```